Bitcode: use 2 bits per symbol
(We used to use 2.5 bits/symbol.)

The decompressor grew from 140 to 148 bytes, and we also need to encode
one more application and a number. However, this all should be swamped
by the gains we get on inputs longer than about a hundred characters,
which is the majority of them.

You can re-submit the solutions anytime with:

    for i in `seq 21`; do stack run -- http-lambdaman-solution-using-bitcoding $i $(cat data/lambdaman/lambdaman${i}.sol) ; sleep 3 ; done